ozflux II updating: 1. updated to the struts2.3.9, spring3.1 and hibernate4.1 2. made changes for css and nav and logo etc

+++ trunk/src/java/au/edu/monash/merc/capture/dao/impl/AuditEventDAO.java
@@ -71,8 +71,7 @@ public class AuditEventDAO extends HibernateGenericDAO<AuditEvent> implements IA
 		Criteria userCriteria = criteria.createCriteria("eventOwner");
 		userCriteria.add(Restrictions.eq("id", uid));
 		criteria.setProjection(Projections.rowCount());
-		// int total = ((Integer) criteria.list().get(0)).intValue();
-		int total = ((Integer) criteria.uniqueResult()).intValue();
+		int total = ((Long) criteria.uniqueResult()).intValue();
 		Pagination<AuditEvent> pev = new Pagination<AuditEvent>(startPageNo, recordsPerPage, total);
 		// query collections by size-per-page
 		Criteria queryCriteria = this.session().createCriteria(this.persistClass);
